sortCommandByAttackerFirstを追加
コマンドソート処理を別関数に切り出した
